Extracting actual changes into extension methods
